Changed accumulator from float to double and added test cases for accuracy check.
git-svn-id: http://alma-dms/svn/sakura/trunk@1263 9a79bd28-2a0b-46f8-beb7-14fc758f186d

+++ libsakura/src/libsakura/sakura.h
@@ -330,22 +330,22 @@ typedef struct {
 	 * @~
 	 * sum of valid data
 	 */
-	float sum;
+	double sum;
 	/**
 	 * @~
 	 * mean of valid data
 	 */
-	float mean;
+	double mean;
 	/**
 	 * @~
 	 * root-mean-square of valid data
 	 */
-	float rms;
+	double rms;
 	/**
 	 * @~
 	 * stddev of valid data
 	 */
-	float stddev;
+	double stddev;
 	/**
 	 * @~
 	 * min value of valid data
